# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OmniCreator AI Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/workspaces:
    get:
      tags: [Workspaces]
      operationId: listWorkspaces
      summary: List workspaces in an organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
    post:
      tags: [Workspaces]
      operationId: createWorkspace
      summary: Create a workspace
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceBody'
      responses:
        '201':
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /workspaces/{workspaceId}:
    get:
      tags: [Workspaces]
      operationId: getWorkspace
      summary: Get workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
    patch:
      tags: [Workspaces]
      operationId: updateWorkspace
      summary: Update workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceBody'
      responses:
        '200':
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'

  # MEMBERS
...
  /templates/{templateId}:
    get:
      tags: [Templates]
      operationId: getTemplate
      summary: Get template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'

  # SOCIAL ACCOUNTS
...
  /organizations/{orgId}/providers:
    get:
      tags: [Providers]
      operationId: listProviders
      summary: List AI provider configurations
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProviderConfig'
    post:
      tags: [Providers]
      operationId: saveProviderConfig
      summary: Save AI provider configuration
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveProviderConfigBody'
      responses:
        '200':
          description: Saved provider config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderConfig'
...
components:
...
  schemas:
...
    Workspace:
      type: object
      required: [id, orgId, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        orgId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        defaultBrandKitId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CreateWorkspaceBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    UpdateWorkspaceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        defaultBrandKitId:
          type: ["string", "null"]

    # MEMBERS
...
    Template:
      type: object
      required: [id, name, type, category, thumbnailUrl, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: ["string", "null"]
        name:
          type: string
        type:
          type: string
        category:
          type: string
        thumbnailUrl:
          type: string
        isSystem:
          type: boolean
        isBrandLocked:
          type: boolean
        createdAt:
          type: string

    # SOCIAL ACCOUNTS
...
    ProviderConfig:
      type: object
      required: [providerKey, configured, taskTypes, createdAt]
      properties:
        providerKey:
          type: string
        name:
          type: string
        configured:
          type: boolean
        taskTypes:
          type: array
          items:
            type: string
        isDefault:
          type: boolean
        priority:
          type: integer
        dailyLimit:
          type: ["integer", "null"]
        health:
          $ref: '#/components/schemas/ProviderHealth'
        createdAt:
          type: string
    ProviderHealth:
      type: object
      required: [configured, available]
      properties:
        configured:
          type: boolean
        available:
          type: boolean
        quotaRemaining:
          type: ["integer", "null"]
        message:
          type: ["string", "null"]
        lastCheckedAt:
          type: ["string", "null"]
    SaveProviderConfigBody:
      type: object
      required: [providerKey]
      properties:
        providerKey:
          type: string
        apiKey:
          type: string
        isDefault:
          type: boolean
        priority:
          type: integer
        dailyLimit:
          type: integer

    # DASHBOARD
...