# packages/api-zod/src/generated/types/providerHealth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OmniCreator AI Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderHealth {
  configured: boolean;
  available: boolean;
  /** @nullable */
  quotaRemaining?: number | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  lastCheckedAt?: string | null;
}